+++ lightninglabs/index.js
@@ -12,7 +12,6 @@ const releaseSwapOutSecret = require('./release_swap_out_secret');
 const serviceSocket = require('./service_socket');
 const subscribeToSwapInStatus = require('./subscribe_to_swap_in_status');
 const subscribeToSwapOutStatus = require('./subscribe_to_swap_out_status');
-const swapInAddress = require('./swap_in_address');
 const swapInFee = require('./swap_in_fee');
 const swapUserId = require('./swap_user_id');
 
@@ -31,7 +30,6 @@ module.exports = {
   serviceSocket,
   subscribeToSwapInStatus,
   subscribeToSwapOutStatus,
-  swapInAddress,
   swapInFee,
   swapUserId,
 };